+++ Core/Source/SolidSyslogCircularBufferPrivate.h
@@ -36,7 +36,7 @@ static inline void CircularBuffer_Report(
     enum SolidSyslogCircularBufferErrors code
 )
 {
-    SolidSyslog_Error(severity, &CircularBufferErrorSource, category, code);
+    SolidSyslog_Error(severity, &CircularBufferErrorSource, category, (int32_t) code);
 }
 
 #endif /* SOLIDSYSLOGCIRCULARBUFFERPRIVATE_H */
