+++ Core/Source/SolidSyslogOriginSdPrivate.h
@@ -41,7 +41,7 @@ static inline void OriginSd_Report(
     enum SolidSyslogOriginSdErrors code
 )
 {
-    SolidSyslog_Error(severity, &OriginSdErrorSource, category, code);
+    SolidSyslog_Error(severity, &OriginSdErrorSource, category, (int32_t) code);
 }
 
 #endif /* SOLIDSYSLOGORIGINSDPRIVATE_H */
