+++ Core/Source/SolidSyslogUdpSenderPrivate.h
@@ -27,7 +27,7 @@ static inline void UdpSender_Report(
     enum SolidSyslogUdpSenderErrors code
 )
 {
-    SolidSyslog_Error(severity, &UdpSenderErrorSource, category, code);
+    SolidSyslog_Error(severity, &UdpSenderErrorSource, category, (int32_t) code);
 }
 
 #endif /* SOLIDSYSLOGUDPSENDERPRIVATE_H */
