+++ Platform/FreeRtos/Source/SolidSyslogFreeRtosMutexPrivate.h
@@ -30,7 +30,7 @@ static inline void FreeRtosMutex_Report(
     enum SolidSyslogFreeRtosMutexErrors code
 )
 {
-    SolidSyslog_Error(severity, &FreeRtosMutexErrorSource, category, code);
+    SolidSyslog_Error(severity, &FreeRtosMutexErrorSource, category, (int32_t) code);
 }
 
 #endif /* SOLIDSYSLOGFREERTOSMUTEXPRIVATE_H */
