+++ Platform/LwipRaw/Source/SolidSyslogLwipRawAddressPrivate.h
@@ -42,7 +42,7 @@ static inline void LwipRawAddress_Report(
     enum SolidSyslogLwipRawAddressErrors code
 )
 {
-    SolidSyslog_Error(severity, &LwipRawAddressErrorSource, category, code);
+    SolidSyslog_Error(severity, &LwipRawAddressErrorSource, category, (int32_t) code);
 }
 
 #endif /* SOLIDSYSLOGLWIPRAWADDRESSPRIVATE_H */
