+++ Platform/MbedTls/Source/SolidSyslogMbedTlsStreamPrivate.h
@@ -28,7 +28,7 @@ static inline void MbedTlsStream_Report(
     enum SolidSyslogMbedTlsStreamErrors code
 )
 {
-    SolidSyslog_Error(severity, &MbedTlsStreamErrorSource, category, code);
+    SolidSyslog_Error(severity, &MbedTlsStreamErrorSource, category, (int32_t) code);
 }
 
 #endif /* SOLIDSYSLOGMBEDTLSSTREAMPRIVATE_H */
