+++ Platform/OpenSsl/Source/SolidSyslogOpenSslAesGcmPolicyPrivate.h
@@ -27,7 +27,7 @@ static inline void OpenSslAesGcmPolicy_Report(
     enum SolidSyslogOpenSslAesGcmPolicyErrors code
 )
 {
-    SolidSyslog_Error(severity, &OpenSslAesGcmPolicyErrorSource, category, code);
+    SolidSyslog_Error(severity, &OpenSslAesGcmPolicyErrorSource, category, (int32_t) code);
 }
 
 #endif /* SOLIDSYSLOGOPENSSLAESGCMPOLICYPRIVATE_H */
