+++ Platform/PlusTcp/Source/SolidSyslogPlusTcpAddressPrivate.h
@@ -37,7 +37,7 @@ static inline void PlusTcpAddress_Report(
     enum SolidSyslogPlusTcpAddressErrors code
 )
 {
-    SolidSyslog_Error(severity, &PlusTcpAddressErrorSource, category, code);
+    SolidSyslog_Error(severity, &PlusTcpAddressErrorSource, category, (int32_t) code);
 }
 
 #endif /* SOLIDSYSLOGPLUSTCPADDRESSPRIVATE_H */
