+++ Platform/PlusTcp/Source/SolidSyslogPlusTcpTcpStreamPrivate.h
@@ -31,7 +31,7 @@ static inline void PlusTcpTcpStream_Report(
     enum SolidSyslogPlusTcpTcpStreamErrors code
 )
 {
-    SolidSyslog_Error(severity, &PlusTcpTcpStreamErrorSource, category, code);
+    SolidSyslog_Error(severity, &PlusTcpTcpStreamErrorSource, category, (int32_t) code);
 }
 
 #endif /* SOLIDSYSLOGPLUSTCPTCPSTREAMPRIVATE_H */
