+++ Platform/Posix/Source/SolidSyslogPosixAddressPrivate.h
@@ -34,7 +34,7 @@ static inline void PosixAddress_Report(
     enum SolidSyslogPosixAddressErrors code
 )
 {
-    SolidSyslog_Error(severity, &PosixAddressErrorSource, category, code);
+    SolidSyslog_Error(severity, &PosixAddressErrorSource, category, (int32_t) code);
 }
 
 #endif /* SOLIDSYSLOGPOSIXADDRESSPRIVATE_H */
