+++ Platform/Windows/Source/SolidSyslogWindowsFilePrivate.h
@@ -23,7 +23,7 @@ static inline void WindowsFile_Report(
     enum SolidSyslogWindowsFileErrors code
 )
 {
-    SolidSyslog_Error(severity, &WindowsFileErrorSource, category, code);
+    SolidSyslog_Error(severity, &WindowsFileErrorSource, category, (int32_t) code);
 }
 
 #endif /* SOLIDSYSLOGWINDOWSFILEPRIVATE_H */
