+++ Platform/Windows/Source/SolidSyslogWinsockResolverPrivate.h
@@ -22,7 +22,7 @@ static inline void WinsockResolver_Report(
     enum SolidSyslogWinsockResolverErrors code
 )
 {
-    SolidSyslog_Error(severity, &WinsockResolverErrorSource, category, code);
+    SolidSyslog_Error(severity, &WinsockResolverErrorSource, category, (int32_t) code);
 }
 
 #endif /* SOLIDSYSLOGWINSOCKRESOLVERPRIVATE_H */
